pyproject.toml
[project]
name = "multibios"
version = "0.1.0"
description = "A Multispecies Bilateral Odor Delivery System"
authors = [
    {name = "Rishika Mohanta",email = "10175857+neurorishika@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "mkinit (>=1.1.0,<2.0.0)",
    "lazy-loader (>=0.4,<0.5)",
    "numpy (>=2.3.2,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "nidaqmx (>=1.2.0,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
